requirements.txt
requests==2.22.0
openpyxl==3.0.2
Faker==2.0.4
pytest==5.4.3
PyMySQL==0.9.3
nb_log==3.4
python_jenkins==1.7.0
ddt==1.2.1
jsonpath==0.82
jenkins==1.0.2
locustio==0.999
paramiko==2.7.2
PyYAML==5.3.1
